+++ alpha_trader/order/__init__.py
@@ -44,7 +44,7 @@ class Order(BaseModel):
     counter_party_name: Union[str, None]
     creation_date: int
     execution_price: Union[float, None]
-    execution_volume: Union[int, None]
+    execution_volume: Union[float, None]
     good_after_date: Union[int, None]
     good_till_date: Union[int, None]
     hourly_change: Union[int, None]
